+++ src/frame.rs
@@ -630,6 +630,7 @@ mod tests {
             quant_table_set_count: Some(2),
             ec: Some(0),
             intra: Some(false),
+            initial_state_delta: None,
         }
     }